[arcane,utils] Déplace une partie de l'implémentation du profilage dans un fichier interne à Arcane.
Cela permettra de l'utiliser dans Arcane sans que les utilisateurs puissent
accéder à l'API interne associée.
