+++ arcane/src/arcane/utils/Profiling.h
@@ -25,29 +25,7 @@
 
 namespace Arcane::impl
 {
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-/*!
- * \brief Statistiques d'exécution d'une boucle.
- */
-struct ARCANE_UTILS_EXPORT ForLoopProfilingStat
-{
- public:
-
-  //! Ajoute les infos de l'exécution \a s
-  void add(const ForLoopOneExecStat& s);
-
-  Int64 nbCall() const { return m_nb_call; }
-  Int64 nbChunk() const { return m_nb_chunk; }
-  Int64 execTime() const { return m_exec_time; }
-
- private:
-
-  Int64 m_nb_call = 0;
-  Int64 m_nb_chunk = 0;
-  Int64 m_exec_time = 0;
-};
+class ForLoopStatInfoListImpl;
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
@@ -70,14 +48,11 @@ class ARCANE_UTILS_EXPORT ScopedStatLoop
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
-
 /*!
- * \brief Statistiques d'exécution d'une boucle.
+ * \brief Statistiques d'exécution des boucles.
  */
 class ARCANE_UTILS_EXPORT ForLoopStatInfoList
 {
-  class Impl;
-
  public:
 
   ForLoopStatInfoList();
@@ -86,11 +61,18 @@ class ARCANE_UTILS_EXPORT ForLoopStatInfoList
  public:
 
   void merge(const ForLoopOneExecStat& loop_stat_info, const ForLoopTraceInfo& loop_trace_info);
-  void print(std::ostream& o);
+
+ public:
+
+  /*!
+   * \internal
+   * \brief Type opaque pour l'implémentation interne.
+   */
+  ForLoopStatInfoListImpl* _internalImpl() { return m_p; }
 
  private:
 
-  Impl* m_p = nullptr;
+  ForLoopStatInfoListImpl* m_p = nullptr;
 };
 
 } // namespace Arcane::impl
